ml/requirements.txt
PyYAML>=6.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
joblib>=1.3.0
pytest>=8.0.0
